+++ packages/db/migrations/0065_nifty_victor_mancha.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "projects" ADD COLUMN "is_token_required" boolean DEFAULT false NOT NULL;